+++ src/backend/x11/compositor/init.rs
@@ -1261,6 +1261,7 @@ impl<C: CompositorConnection> Compositor<C> {
             recording_pbo: [None, None],
             recording_cursor: [None, None],
             recording_cursor_sampler: None,
+            recording_last_cursor: None,
             recording_frame_region: [(0, 0, screen_w, screen_h); 2],
             recording_region: (0, 0, screen_w, screen_h),
             recording_output_size: (screen_w, screen_h),